#799 Public project visibility
This MR contains the next portion of fixes:

1. Public projects are not visible - there was incorrect generation of public projects list
1. Partially covers issue mlreef#797

The point of the changes in the MR is the next:

Endpoints that return projects set (/api/v1/projects, /api/v1/data-projects, /api/v1/code-projects) can work in 2 modes now. They returns Pageble and nonpageable results. It can be achieved by using 'page', 'size', 'sort' params in request urls. If these parameters are presented the paged result will be returned. If not then result will be wrapped by a simple array.

There are some detail if implementation:

1. nonpaged result can contain 2000 records max. It is set implicitly. It protects Backend from a performance crash. To get 2000+ items in response Paged request must be used
1. Paged results are also have the limits in 2000 items max.
1. If only 'size=2000' param is presented in request then it is supposed as nonpaged result was requested.
1. Nonpaged requests are deprecated and were left for backward compatibility only, not to break current FE
1. To get result wrapped by Page object 'page', 'size', 'sort' - any of these parameters should be presented in url

relates to: mlreef#797

closes: mlreef#799